fix(mvp5): architecture Port migration (ADR-052) + ArchTest gap + BUG-M5-002/003/009/010 (#3)
* refactor(arch): ADR-052 hexagonal Port migration + ArchTest coverage gap fix

Extract 5 Port interfaces in common.spi to remove direct cross-module
repository/service dependencies (BUG-M5-009 + migration plan P1):

- AirQualityPort (esg → environment.repository) — fix BUG-M5-009
- EnvironmentBroadcastPort + SseBroadcastPort (scheduler → environment/notification)
- AlertPort + NotificationPort (safety → alert/notification)
- BillingInvoiceGeneratedEvent + @TransactionalEventListener (billing → audit)

Each Port has a provider-side @Component adapter; consumers inject the Port.
Add 4 new ModuleBoundaryArchTest rules (safety/billing/scheduler/bms boundary)
to close the coverage gap that let BUG-M5-009 slip through (73 → 77 rules).

Verified: ModuleBoundaryArchTest 77/77 PASS; all migrated unit tests PASS;
no regression (remaining failures are pre-existing ApplicationContext-load
issues, unrelated to Port beans).

* fix(arch): BUG-M5-010 bean-name conflicts blocking ApplicationContext (366 false-fails)

Two pairs of classes shared a simple-name across modules; @RestController/@Service
with no explicit bean name produced ConflictingBeanDefinitionException that blocked
the entire Spring context startup.

- InvoiceController: billing.api (admin-ops) vs citizen.api (citizen self-service)
  — different audience/aggregate/path/role → explicit bean names
  (billingInvoiceController / citizenInvoiceController), no @Qualifier needed
  (controllers route by path).
- AuditLogService: audit.service (M5-4 active, table audit.events) vs
  common.service (dead code — logAction has 0 callers; repo/entity only
  referenced each other) → deleted the common chain (service + repo + entity
  + 2 tests). Orphan table public.audit_log (V15) left in place.

Verified: ConflictingBeanDefinitionException 0 remaining instances.

* fix(billing): BUG-M5-002 invoice NPE + BUG-M5-003 reconciliation mock + carry-over compile fixes

- BUG-M5-002 (InvoiceGenerationServiceTest NPE): mock invoiceRepository.save()
  returned null + skipped @PrePersist. Added @BeforeEach stub that mimics
  JPA save (assigns id + generatedAt). 6/6 PASS.
- BUG-M5-003 (BillingReconciliationServiceTest wrong figures): mockRawEventCost()
  helper ignored its cost arg (always returned List.of()). Now builds a real
  MeteringEvent with costUsdCents. 5/5 PASS.
- billing→audit decouple (migration C4): InvoiceGenerationService publishes
  BillingInvoiceGeneratedEvent instead of injecting AuditLogService.
- Carry-over compile fixes for MVP5 session (MeteringEvent, RoiCalculation,
  EsgMetricRepository, TenantContext, TenantService, KafkaProducerService,
  Flink NgsiLdMessage + FlinkMultiTenantConcurrentIT).

* docs(mvp5): ADR-052 + cross-module migration plan + re-test report + CLAUDE.md checklist

- ADR-052: Hexagonal Port for cross-module dependency — architectural standard
  (Port in common.spi neutral package, provider adapter, consumer injects Port).
- mvp5-cross-module-migration-plan.md: full-system audit of 28 cross-module
  couplings, classified B/D/C/A; ArchTest coverage gap identified; P1 migrated.
- mvp5-sprint1-4-retest-20260630.md: re-test of 4 sprints, BUG-M5-002/003/004/
  008/009/010 findings + fixes; BUG-M5-004 (LOTUS scoring) escalated to BA.
- mvp5-sprint-plan.md: sign-off updated (BUG-M5-009/010 fixed, P1 migration done).
- CLAUDE.md: cross-module dependency rule + SA Review Checklist item 11
  (verify ModuleBoundaryArchTest PASS before DONE on any cross-module import).

+++ backend/src/main/java/com/uip/backend/billing/service/RoiCalculationService.java
@@ -219,7 +219,7 @@ public class RoiCalculationService {
 
         // CO2 savings from energy reduction
         // Assume UIP automation reduces energy consumption by 15% (configurable)
-        long energySavedKwh = roiConfig.getAvgEnergyConsumptionKwh()
+        long energySavedKwh = BigDecimal.valueOf(roiConfig.getAvgEnergyConsumptionKwh())
                 .multiply(roiConfig.getEnergySavingsFactor())
                 .longValue();
